requirements.txt
alembic==0.9.2
amqp==2.1.4
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
cssmin==0.2.0
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Assets==0.12
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jsmin==2.2.2
kombu==4.0.2
Mako==1.0.6
MarkupSafe==1.0
pkg-resources==0.0.0
python-dateutil==2.6.0
python-editor==1.0.3
python-valve==0.1.1
pytz==2017.2
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.10
urllib3==1.21.1
vine==1.1.3
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.12.2
WTForms==2.1
